+++ packages/remix-deno/index.ts
@@ -14,11 +14,13 @@ export {
 } from "./implementations.ts";
 
 export {
+  broadcastDevReady,
   createSession,
   defer,
   isCookie,
   isSession,
   json,
+  logDevReady,
   MaxPartSizeExceededError,
   redirect,
   unstable_composeUploadHandlers,