+++ packages/remix-dev/cli/commands.ts
@@ -1,26 +1,31 @@
 import * as path from "path";
 import { execSync } from "child_process";
+import inspector from "inspector";
 import * as fse from "fs-extra";
+import getPort, { makeRange } from "get-port";
 import ora from "ora";
 import prettyMs from "pretty-ms";
 import * as esbuild from "esbuild";
 import NPMCliPackageJson from "@npmcli/package-json";
+import { coerce } from "semver";
 
 import * as colors from "../colors";
 import * as compiler from "../compiler";
 import * as devServer from "../devServer";
-import * as devServer2 from "../devServer2";
+import * as devServer_unstable from "../devServer_unstable";
 import type { RemixConfig } from "../config";
 import { readConfig } from "../config";
 import { formatRoutes, RoutesFormat, isRoutesFormat } from "../config/format";
-import { log } from "../logging";
 import { createApp } from "./create";
-import { getPreferredPackageManager } from "./getPreferredPackageManager";
+import { detectPackageManager } from "./detectPackageManager";
 import { setupRemix, isSetupPlatform, SetupPlatform } from "./setup";
 import runCodemod from "../codemod";
 import { CodemodError } from "../codemod/utils/error";
 import { TaskError } from "../codemod/utils/task";
 import { transpile as convertFileToJS } from "./useJavascript";
+import { warnOnce } from "../warnOnce";
+import type { Options } from "../compiler/options";
+import { getAppDependencies } from "../dependencies";
 
 export async function create({
   appTemplate,
@@ -79,7 +84,7 @@ export async function init(
 
   let initPackageJson = path.resolve(initScriptDir, "package.json");
   let isTypeScript = fse.existsSync(path.join(projectDir, "tsconfig.json"));
-  let packageManager = getPreferredPackageManager();
+  let packageManager = detectPackageManager() ?? "npm";
 
   if (await fse.pathExists(initPackageJson)) {
     execSync(`${packageManager} install`, {
@@ -126,7 +131,7 @@ export async function setup(platformArg?: string) {
 
   await setupRemix(platform);
 
-  log(`Successfully setup Remix for ${platform}.`);
+  console.log(`Successfully setup Remix for ${platform}.`);
 }
 
 export async function routes(
@@ -145,9 +150,9 @@ export async function build(
   modeArg?: string,
   sourcemap: boolean = false
 ): Promise<void> {
-  let mode = compiler.parseMode(modeArg ?? "", "production");
+  let mode = parseMode(modeArg) ?? "production";
 
-  log(`Building Remix app in ${mode} mode...`);
+  console.log(`Building Remix app in ${mode} mode...`);
 
   if (modeArg === "production" && sourcemap) {
     console.warn(
@@ -166,24 +171,36 @@ export async function build(
 
   let start = Date.now();
   let config = await readConfig(remixRoot);
-  fse.emptyDirSync(config.assetsBuildDirectory);
-  await compiler.build(config, {
+  let options: Options = {
     mode,
     sourcemap,
-    onCompileFailure: (failure) => {
-      compiler.logCompileFailure(failure);
-      throw Error();
-    },
+    onWarning: warnOnce,
+  };
+  if (mode === "development" && config.future.unstable_dev) {
+    let dev = await resolveDevBuild(config);
+    options.devHttpOrigin = {
+      scheme: dev.httpScheme,
+      host: dev.httpHost,
+      port: dev.httpPort,
+    };
+    options.devWebsocketPort = dev.websocketPort;
+  }
+
+  fse.emptyDirSync(config.assetsBuildDirectory);
+  await compiler.build({ config, options }).catch((thrown) => {
+    compiler.logThrown(thrown);
+    process.exit(1);
   });
 
-  log(`Built in ${prettyMs(Date.now() - start)}`);
+  console.log(`Built in ${prettyMs(Date.now() - start)}`);
 }
 
+// TODO: replace watch in v2
 export async function watch(
   remixRootOrConfig: string | RemixConfig,
   modeArg?: string
 ): Promise<void> {
-  let mode = compiler.parseMode(modeArg ?? "", "development");
+  let mode = parseMode(modeArg) ?? "development";
   console.log(`Watching Remix app in ${mode} mode...`);
 
   let config =
@@ -191,29 +208,41 @@ export async function watch(
       ? remixRootOrConfig
       : await readConfig(remixRootOrConfig);
 
-  devServer.liveReload(config, {
-    mode,
-    onInitialBuild: (durationMs) =>
-      console.log(`💿 Built in ${prettyMs(durationMs)}`),
-  });
+  devServer.liveReload(config);
   return await new Promise(() => {});
 }
 
 export async function dev(
   remixRoot: string,
-  modeArg?: string,
-  flags: { port?: number; appServerPort?: number } = {}
-) {
-  let config = await readConfig(remixRoot);
-  let mode = compiler.parseMode(modeArg ?? "", "development");
+  flags: {
+    debug?: boolean;
+    port?: number; // TODO: remove for v2
 
-  if (config.future.unstable_dev !== false) {
-    await devServer2.serve(config, flags);
+    // unstable_dev
+    command?: string;
+    httpScheme?: string;
+    httpHost?: string;
+    httpPort?: number;
+    restart?: boolean;
+    websocketPort?: number;
+  } = {}
+) {
+  if (process.env.NODE_ENV && process.env.NODE_ENV !== "development") {
+    console.warn(
+      `Forcing NODE_ENV to be 'development'. Was: ${process.env.NODE_ENV}`
+    );
+  }
+  process.env.NODE_ENV = "development";
+  if (flags.debug) inspector.open();
+
+  let config = await readConfig(remixRoot);
+
+  if (config.future.unstable_dev === false) {
+    await devServer.serve(config, flags.port);
     return await new Promise(() => {});
   }
 
-  await devServer.serve(config, mode, flags.port);
-  return await new Promise(() => {});
+  await devServer_unstable.serve(config, await resolveDevServe(config, flags));
 }
 
 export async function codemod(
@@ -257,9 +286,16 @@ let entries = ["entry.client", "entry.server"];
 
 // @ts-expect-error available in node 12+
 // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Intl/ListFormat#browser_compatibility
-let listFormat = new Intl.ListFormat("en", {
+let conjunctionListFormat = new Intl.ListFormat("en", {
   style: "long",
   type: "conjunction",
+});
+
+// @ts-expect-error available in node 12+
+// https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Intl/ListFormat#browser_compatibility
+let disjunctionListFormat = new Intl.ListFormat("en", {
+  style: "long",
+  type: "disjunction",
 });
 
 export async function generateEntry(
@@ -278,7 +314,7 @@ export async function generateEntry(
 
   if (!entries.includes(entry)) {
     let entriesArray = Array.from(entries);
-    let list = listFormat.format(entriesArray);
+    let list = conjunctionListFormat.format(entriesArray);
 
     console.error(
       colors.error(`Invalid entry file. Valid entry files are ${list}`)
@@ -288,6 +324,20 @@ export async function generateEntry(
 
   let pkgJson = await NPMCliPackageJson.load(config.rootDirectory);
   let deps = pkgJson.content.dependencies ?? {};
+
+  let maybeReactVersion = coerce(deps.react);
+  if (!maybeReactVersion) {
+    let react = ["react", "react-dom"];
+    let list = conjunctionListFormat.format(react);
+    throw new Error(
+      `Could not determine React version. Please install the following packages: ${list}`
+    );
+  }
+
+  let type =
+    maybeReactVersion.major >= 18 || maybeReactVersion.raw === "0.0.0"
+      ? ("stream" as const)
+      : ("string" as const);
 
   let serverRuntime = deps["@remix-run/deno"]
     ? "deno"
@@ -303,7 +353,7 @@ export async function generateEntry(
       "@remix-run/cloudflare",
       "@remix-run/node",
     ];
-    let formattedList = listFormat.format(serverRuntimes);
+    let formattedList = disjunctionListFormat.format(serverRuntimes);
     console.error(
       colors.error(
         `Could not determine server runtime. Please install one of the following: ${formattedList}`
@@ -312,9 +362,9 @@ export async function generateEntry(
     return;
   }
 
-  let clientRuntime = deps["@remix-run/react"] ? "react" : undefined;
+  let clientRenderer = deps["@remix-run/react"] ? "react" : undefined;
 
-  if (!clientRuntime) {
+  if (!clientRenderer) {
     console.error(
       colors.error(
         `Could not determine runtime. Please install the following: @remix-run/react`
@@ -326,11 +376,12 @@ export async function generateEntry(
   let defaultsDirectory = path.resolve(__dirname, "..", "config", "defaults");
   let defaultEntryClient = path.resolve(
     defaultsDirectory,
-    `entry.client.${clientRuntime}.tsx`
+    `entry.client.${clientRenderer}-${type}.tsx`
   );
   let defaultEntryServer = path.resolve(
     defaultsDirectory,
-    `entry.server.${serverRuntime}.tsx`
+    serverRuntime,
+    `entry.server.${clientRenderer}-${type}.tsx`
   );
 
   let isServerEntry = entry === "entry.server";
@@ -406,3 +457,110 @@ async function createClientEntry(
   let contents = await fse.readFile(inputFile, "utf-8");
   return contents;
 }
+
+let parseMode = (
+  mode?: string
+): compiler.CompileOptions["mode"] | undefined => {
+  if (mode === undefined) return undefined;
+  if (mode === "development") return mode;
+  if (mode === "production") return mode;
+  if (mode === "test") return mode;
+  console.error(`Unrecognized mode: ${mode}`);
+  process.exit(1);
+};
+
+let findPort = async () => getPort({ port: makeRange(3001, 3100) });
+
+type DevBuildFlags = {
+  httpScheme: string;
+  httpHost: string;
+  httpPort: number;
+  websocketPort: number;
+};
+let resolveDevBuild = async (
+  config: RemixConfig,
+  flags: Partial<DevBuildFlags> = {}
+): Promise<DevBuildFlags> => {
+  let dev = config.future.unstable_dev;
+  if (dev === false) throw Error("This should never happen");
+
+  // prettier-ignore
+  let httpScheme =
+    flags.httpScheme ??
+    (dev === true ? undefined : dev.httpScheme) ??
+    "http";
+  // prettier-ignore
+  let httpHost =
+    flags.httpHost ??
+    (dev === true ? undefined : dev.httpHost) ??
+    "localhost";
+  // prettier-ignore
+  let httpPort =
+    flags.httpPort ??
+    (dev === true ? undefined : dev.httpPort) ??
+    (await findPort());
+  // prettier-ignore
+  let websocketPort =
+    flags.websocketPort ??
+    (dev === true ? undefined : dev.websocketPort) ??
+    (await findPort());
+
+  return {
+    httpScheme,
+    httpHost,
+    httpPort,
+    websocketPort,
+  };
+};
+
+type DevServeFlags = DevBuildFlags & {
+  command: string;
+  restart: boolean;
+};
+let resolveDevServe = async (
+  config: RemixConfig,
+  flags: Partial<DevServeFlags> = {}
+): Promise<DevServeFlags> => {
+  let dev = config.future.unstable_dev;
+  if (dev === false) throw Error("Cannot resolve dev options");
+
+  let { httpScheme, httpHost, httpPort, websocketPort } = await resolveDevBuild(
+    config,
+    flags
+  );
+
+  // prettier-ignore
+  let command =
+    flags.command ??
+    (dev === true ? undefined : dev.command)
+  if (!command) {
+    command = `remix-serve ${path.relative(
+      process.cwd(),
+      config.serverBuildPath
+    )}`;
+
+    let usingRemixAppServer =
+      getAppDependencies(config, true)["@remix-run/serve"] !== undefined;
+    if (!usingRemixAppServer) {
+      console.error(
+        [
+          `Remix dev server command defaulted to '${command}', but @remix-run/serve is not installed.`,
+          "If you are using another server, specify how to run it with `-c` or `--command` flag.",
+          "For example, `remix dev -c 'node ./server.js'`",
+        ].join("\n")
+      );
+      process.exit(1);
+    }
+  }
+  let restart =
+    flags.restart ?? (dev === true ? undefined : dev.restart) ?? true;
+
+  return {
+    command,
+    httpScheme,
+    httpHost,
+    httpPort,
+    websocketPort,
+    restart,
+  };
+};