+++ packages/remix-react/index.tsx
@@ -1,6 +1,8 @@
 export type { RemixBrowserProps } from "./browser";
 export { RemixBrowser } from "./browser";
 export type {
+  FormEncType,
+  FormMethod,
   FormProps,
   Location,
   NavigateFunction,
@@ -62,8 +64,6 @@ export {
   useActionData,
   RemixContext as UNSAFE_RemixContext,
 } from "./components";
-
-export type { FormMethod, FormEncType } from "./data";
 
 export type { ThrownResponse } from "./errors";
 export { useCatch } from "./errorBoundaries";